refactor: move FullPipelineContext and spawn_two_phase_async to runtime.rs
Move ~420 lines of pipeline orchestration from main.rs to runtime.rs,
alongside ChainRuntime and CommonChannels. Replace separate catch_up_only,
repair, repair_scope fields with a single IndexerMode field.
